use float to get sensor temp value

+++ src/Monitor/ThermalZone.cs
@@ -1,6 +1,6 @@
 namespace AvaloniaInside.SystemManager.Monitor;
 
-public class ThermalZone : IntervalCollection<int>
+public class ThermalZone : IntervalCollection<float>
 {
     private readonly string _path;
 
@@ -9,8 +9,8 @@ public class ThermalZone : IntervalCollection<int>
         _path = path;
     }
 
-    protected override async Task<int> NewValueAsync(CancellationToken cancellationToken) =>
-        int.Parse((await File.ReadAllTextAsync(_path, cancellationToken)).Trim()) / 100;
+    protected override async Task<float> NewValueAsync(CancellationToken cancellationToken) =>
+        float.Parse((await File.ReadAllTextAsync(_path, cancellationToken)).Trim()) / 100f;
     
     public Task<string> GetTypeAsync(CancellationToken cancellationToken) =>
         File.ReadAllTextAsync(_path, cancellationToken);
